movies-python-bolt/requirements.txt
Flask==0.10.1
Flask-Cors==3.0.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
neo4j-driver==1.4.0
neotime==1.0.0
pytz==2018.4
six==1.11.0
Werkzeug==0.14.1
